Use random access while analysis
Changes the reading behavior to use FileChannel.transferTo while analyzing instead of resetting the position of the FileChannel backing the Okio buffered source.

This change requires updating the indexed data to know not only the position, but also the size of every record. That way we transfer exactly only what we need.

Benchmarking on a Pixel 3, airplane mode, with 10 iterations on a large heap dump (from ProfiledTest):

Before: Durations (ms): median 11884.5, min 11802, max 12021
After: Durations (ms): median 8281.5, min 8181, max 8417

That's a 30% improvement!

Related change: updated the minimum required Okio release to 1.14 because that's when Buffer started implementing WritableByteChannel.

Fixes #1876